Function code. Reads in fake G4 output file and sucessfully digitizes. Added functionality for -Peak amplitude         *tested and works -Timing offset between bars (ie. cable delay)         *needs testing -Hard coded for all events to start at 0
Time-walk effect seen in output data, among other expected correlations (peak amp is proportional to adc_int, etc.).

+++ src/g4sbs_data.cxx
@@ -178,10 +178,10 @@ namespace TSBSGeant4 {
   {
     if(!tree)return(false);
     chan = new std::vector<int>;
-    adc = new std::vector<int>;
-    tdc_l = new std::vector<int>;
-    tdc_t = new std::vector<int>;
-    amp = new std::vector<int>;
+    adc = new std::vector<double>;
+    tdc_l = new std::vector<double>;
+    tdc_t = new std::vector<double>;
+    amp = new std::vector<double>;
     
     b_nchan = tree->Branch(Form("%s.nchan", prefix), &nchan);
     b_chan = tree->Branch(Form("%s.chan", prefix), &chan);
